Fix typo and optimise imports
